# backend/outputs/ApiContract.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PIA Developer Console API",
    "description": "PIA Platform Observability System \u2014 All data served through stable projections and versioned contracts.",
    "version": "2.0.0",
    "x-stability": "frozen"
  },
  "paths": {
...
    "/api/v1/benchmark": {
      "post": {
        "tags": [
          "benchmark"
        ],
        "summary": "Trigger Benchmark",
        "operationId": "trigger_benchmark_api_v1_benchmark_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BenchmarkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BenchmarkJobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/benchmark/{job_id}/progress": {
      "get": {
        "tags": [
          "benchmark"
        ],
        "summary": "Get Benchmark Progress",
        "operationId": "get_benchmark_progress_api_v1_benchmark__job_id__progress_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BenchmarkProgressResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/benchmark/{job_id}/result": {
      "get": {
        "tags": [
          "benchmark"
        ],
        "summary": "Get Benchmark Result",
        "operationId": "get_benchmark_result_api_v1_benchmark__job_id__result_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BenchmarkResultResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BenchmarkJobResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "status"
        ],
        "title": "BenchmarkJobResponse"
      },
      "BenchmarkProgressResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "progress": {
            "type": "integer",
            "title": "Progress"
          },
          "metrics": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metrics"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "status",
          "progress"
        ],
        "title": "BenchmarkProgressResponse"
      },
      "BenchmarkRequest": {
        "properties": {
          "dataset": {
            "type": "string",
            "title": "Dataset"
          }
        },
        "type": "object",
        "required": [
          "dataset"
        ],
        "title": "BenchmarkRequest"
      },
      "BenchmarkResultResponse": {
        "properties": {
          "metrics": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metrics"
          }
        },
        "type": "object",
        "required": [
          "metrics"
        ],
        "title": "BenchmarkResultResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}